+++ src/test/java/com/woowahan/lotto/LottoControllerTest.java
@@ -17,34 +17,9 @@ public class LottoControllerTest {
     }
 
     @Test
-    public void 랜덤_숫자_사이즈_확인() {
-        Lotto lotto = new Lotto();
-
-        assertEquals(6, lotto.getLotto().size());
-    }
-
-    @Test
     public void 입력금액만큼_로또생성() {
         List<Lotto> totalLotto = lottoController.buy(14000);
         assertEquals(14, totalLotto.size());
-    }
-
-    @Test
-    public void 로또번호_오름차순으로_정렬() {
-        Lotto lotto = new Lotto();
-
-        List<Integer> sortedList = lotto.sort(lotto.getLotto());
-
-        boolean isSorted = true;
-
-        for (int i = 0; i < sortedList.size() - 1; i++) {
-            if (sortedList.get(i) > sortedList.get(i + 1)) {
-                isSorted = false;
-                break;
-            }
-        }
-
-        assertTrue(isSorted);
     }
 
     @Test
@@ -69,12 +44,12 @@ public class LottoControllerTest {
                 new Lotto(Arrays.asList(1, 3, 4, 9, 13, 20))
         );
         List<Integer> lucky = Arrays.asList(1, 2, 3, 4, 5, 6);
-        Map<Integer, Integer> luckyInfo = lottoController.getWinInfo(lottos, lucky);
-        Map<Integer, Integer> expected = new HashMap<>();
-        expected.put(5000, 1);
-        expected.put(50000, 1);
-        expected.put(1500000, 0);
-        expected.put(2000000000, 0);
+        Map<PriceInfo, Integer> luckyInfo = lottoController.checkAllLottos(lottos, lucky);
+        Map<PriceInfo, Integer> expected = new HashMap<>();
+        expected.put(PriceInfo.THREE, 1);
+        expected.put(PriceInfo.FOUR, 1);
+        expected.put(PriceInfo.FIVE, 0);
+        expected.put(PriceInfo.SIX, 0);
         assertEquals(expected, luckyInfo);
     }
